requirements_frozen.txt
bashlex==0.15
boto==2.49.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
compiledb==0.10.1
Deprecated==1.2.12
distro==1.5.0
downloadutil==1.0.2
idna==2.10
mypy==0.812
mypy-extensions==0.4.3
overrides==6.1.0
psutil==5.8.0
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
semantic-version==2.8.5
shutilwhich==1.1.0
six==1.16.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
typing-utils==0.0.3
urllib3==1.26.4
wrapt==1.12.1
yugabyte-pycommon==1.9.15
